Changed wrong type in LiteralNode from 'double' to 'float'

+++ include/parser/ast/expressions/literal_node.hpp
@@ -11,14 +11,14 @@
 ///     Represents literal values
 class LiteralNode : public ExpressionNode {
   public:
-    explicit LiteralNode(std::variant<int, double, std::string, bool, char> &value, const std::string &type) :
+    explicit LiteralNode(std::variant<int, float, std::string, bool, char> &value, const std::string &type) :
         value(std::move(value)) {
         this->type = type;
     }
 
     /// value
     ///     the literal value
-    std::variant<int, double, std::string, bool, char> value;
+    std::variant<int, float, std::string, bool, char> value;
 };
 
 #endif
